Updates
Added functionality to perform an instance count on the different rules
in R0.
This allows a better calculation as to what paths are executed in the
machine.
Consider a rule R5, Used twice, once in R6 and another in R7.
This will have a usage of "2" but R6 may be used 9 times in R12, and R12
used again twice in R0.  This would mean R5 is actually used 36 times,
not including the usage from R7.